Fix event-sourced FSM example #50
- Fix bug with unawaited Become in event handlers by using delayed seyip
via additional State property. In general, Become should not be used
during replay, we're only interested in a final state (behavior)
- Use delayed behavior setup in base command handler. Once command is
handled and events are generated, we can just call Become, since event
handler will only set State. That makes event handlers non-async and
fixed aforementioned bug
- Introduce ctor overload to set initial behavior via same mechanics of
delayed setup. Done in OnActivate
- Introduce helper OnReceive method in base class to simplify command
handler signatures
- Collapse separate files for consistency with the rest of examples
- Use newly added Behavior.Fire to simplify cmd/query handling
- Use newly added overload of Behavior.Become(string)
- Polished console app experience and added aditional instructions for
running it with GES

Huge thanks to @romanstetsenko for this initial sample, which
highlighted some of the shortcomings of current implementation and gave
me pointers how to improve it and make this example so bloody clean :)