Request origin #111
- It is useful to differentiate between unhandled requests sent
externally or from a background behavior

Example: background behavior Fires message, which is sent
via Self. There could be already request in a queue which changes the
current behavior just before message sent by background behavior arrives (races).

The message sent from the backround behavior now may not find a handler,
but that's ok and we want to silently ignore that instead of throwing
exception back.